Move Linux-only input pieces into jfn-linux-util
jfn-input had cfg-gated keysym/xkb modules and the raw-key dispatch path.
Relocate them to jfn-linux-util (already #![cfg(target_os = "linux")]),
leaving jfn-input platform-neutral and cfg-free.
