+++ src/wayland/src/input_lifecycle.rs
@@ -16,10 +16,10 @@ use crate::input::{Callbacks, JfnInputWayland};
 use jfn_platform_abi::cursor::CursorShape;
 
 use jfn_input::{
-    jfn_input_dispatch_char, jfn_input_dispatch_history_nav, jfn_input_dispatch_key_raw,
-    jfn_input_dispatch_keyboard_focus, jfn_input_dispatch_mouse_button,
-    jfn_input_dispatch_mouse_move, jfn_input_dispatch_scroll,
+    jfn_input_dispatch_char, jfn_input_dispatch_history_nav, jfn_input_dispatch_keyboard_focus,
+    jfn_input_dispatch_mouse_button, jfn_input_dispatch_mouse_move, jfn_input_dispatch_scroll,
 };
+use jfn_linux_util::input::jfn_input_dispatch_key_raw;
 
 const CALLBACKS: Callbacks = Callbacks {
     mouse_move: Some(jfn_input_dispatch_mouse_move),
